Beginning to write a STL

+++ src/common/types.h
@@ -11,7 +11,7 @@ typedef int int32_t;
 typedef unsigned int uint32_t;
 typedef long long int int64_t;
 typedef unsigned long long int uint64_t;
-typedef uint32_t size_t;
+typedef unsigned long int size_t;
 
 typedef uint32_t uintptr_t;
